% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{escpos}
\alias{escpos}
\title{Create an \code{escpos} object for accumulating print commands}
\usage{
escpos(host, port = 9100L)
}
\description{
Create an \code{escpos} object for accumulating print commands
}
\examples{
escpos(epson_ip) |>
  pos_plaintext(
    stringi::stri_rand_lipsum(1)
  ) |>
  pos_lf(2) |>
  pos_plaintext(
    paste0(capture.output(
      str(mtcars, width = 40, strict.width = "cut")
    ), collapse = "\n")
  ) |>
  pos_lf(2L) |>
  pos_plaintext(
    stringi::stri_rand_lipsum(1, start_lipsum = FALSE)
  ) |>
  pos_lf(3) |>
  pos_cut() |>
  pos_print()
}
