% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpos-package.R, R/png_to_raster.R
\docType{package}
\name{ggpos}
\alias{ggpos}
\alias{ggpos-package}
\title{Print {ggplot2} or {grid} Objects to ESC/POS Compatible Network Devices}
\usage{
ggpos(
  plot = ggplot2::last_plot(),
  host_pos,
  port = 9100L,
  scale = 2,
  width = 280,
  height = 280,
  units = "px",
  dpi = 144,
  bg = "white",
  ...
)
}
\arguments{
\item{plot}{Plot to save, defaults to last plot displayed.}

\item{host_pos}{hostname or IP address of the ESC/POS compatible network device}

\item{port}{port the ESC/POS compatible device is listening on; defaults to \code{9100L}}

\item{scale, width, height, units, dpi, bg}{same as their \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} counterparts but with
sensible defaults for ESC/POS devices.}

\item{...}{other parameters to pass on to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}
}
\description{
ESC/POS devices, such as receipt printers, have the ability to print their own form of
raster bitmap graphics. Tools are provided to turn {ggplot2} or {grid} objects into such raster
bitmaps and print them to ESC/POS compatible devices.

ESC/POS printers do not have a standard width and some have higher resolutions than others.
These defaults work well for a small, square plot on my used
\href{https://www.epson.com.au/pos/products/receiptprinters/DisplaySpecs.asp?id=tmt88v}{Epson TM-T88V}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/ggpos}
  \item Report bugs at \url{https://github.com/hrbrmstr/ggpos/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
