% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_to_raster.R
\name{png_to_raster}
\alias{png_to_raster}
\title{Convert any png file to ESC/POS raster format}
\usage{
png_to_raster(png_file, color = FALSE)
}
\arguments{
\item{png_file}{path to PNG file}

\item{color}{if \code{TRUE}, an attempt will be made to dither the result}
}
\value{
path to a temporary file in ESC/POS raster bitmap format or \code{""} if an error occurred
}
\description{
Convert any png file to ESC/POS raster format
}
