% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{pos_align}
\alias{pos_align}
\title{Text alignment}
\usage{
pos_align(pos_obj, set = c("left", "right", "center"))
}
\arguments{
\item{pos_obj}{object created with \code{\link[=escpos]{escpos()}}}

\item{set}{one of "\code{left}", "\code{right}", "\code{center}"}
}
\value{
\code{pos_obj} (invisibly)
}
\description{
Text alignment
}
