% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{pos_bold}
\alias{pos_bold}
\title{Bold text}
\usage{
pos_bold(pos_obj, set = c("on", "off"))
}
\arguments{
\item{pos_obj}{object created with \code{\link[=escpos]{escpos()}}}

\item{set}{one of "\code{on}", "\code{off}"}
}
\value{
\code{pos_obj} (invisibly)
}
\description{
Bold text
}
