% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{pos_color}
\alias{pos_color}
\title{Set color}
\usage{
pos_color(pos_obj, set = c("one", "two"))
}
\arguments{
\item{pos_obj}{object created with \code{\link[=escpos]{escpos()}}}

\item{set}{one of "\code{one}", "\code{two}"}
}
\value{
\code{pos_obj} (invisibly)
}
\description{
Set color
}
