% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{pos_cut}
\alias{pos_cut}
\title{Issue a partial or full cut command to the printer}
\usage{
pos_cut(pos_obj, mode = c("part", "full"), feed_lines = 5L)
}
\arguments{
\item{pos_obj}{object created with \code{\link[=escpos]{escpos()}}}

\item{mode}{one of \code{part} or \code{full}}

\item{feed_lines}{how many line feeds to issue before the cut}
}
\description{
Issue a partial or full cut command to the printer
}
\examples{
escpos(epson_ip) |>
  pos_plaintext(
    stringi::stri_rand_lipsum(1)
  ) |>
  pos_lf(2) |>
  pos_plaintext(
    paste0(capture.output(
      str(mtcars, width = 40, strict.width = "cut")
    ), collapse = "\n")
  ) |>
  pos_lf(2L) |>
  pos_plaintext(
    stringi::stri_rand_lipsum(1, start_lipsum = FALSE)
  ) |>
  pos_lf(3) |>
  pos_cut() |>
  pos_print()
}
