% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{pos_font}
\alias{pos_font}
\title{Font choice}
\usage{
pos_font(pos_obj, set = c("a", "b", "c"))
}
\arguments{
\item{pos_obj}{object created with \code{\link[=escpos]{escpos()}}}

\item{set}{one of "\code{a}", "\code{b}", "\code{c}"}
}
\value{
\code{pos_obj} (invisibly)
}
\description{
Font choice
}
