% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{pos_inverted}
\alias{pos_inverted}
\title{Invert printing}
\usage{
pos_inverted(pos_obj, set = c("true", "false"))
}
\arguments{
\item{pos_obj}{object created with \code{\link[=escpos]{escpos()}}}

\item{set}{one of "\code{true}", "\code{false}"}
}
\value{
\code{pos_obj} (invisibly)
}
\description{
Invert printing
}
