% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{pos_plaintext}
\alias{pos_plaintext}
\title{Send plaintext to the printer}
\usage{
pos_plaintext(pos_obj, txt)
}
\arguments{
\item{pos_obj}{object created with \code{\link[=escpos]{escpos()}}}

\item{txt}{character vector. If length > 1, each will be sent to the printer}
}
\description{
Send plaintext to the printer
}
\examples{
escpos(epson_ip) |>
  pos_plaintext(
    stringi::stri_rand_lipsum(1)
  ) |>
  pos_lf(2) |>
  pos_plaintext(
    paste0(capture.output(
      str(mtcars, width = 40, strict.width = "cut")
    ), collapse = "\n")
  ) |>
  pos_lf(2L) |>
  pos_plaintext(
    stringi::stri_rand_lipsum(1, start_lipsum = FALSE)
  ) |>
  pos_lf(3) |>
  pos_cut() |>
  pos_print()
}
