% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{pos_plot}
\alias{pos_plot}
\title{Turn a ggplot (or other grid object) intto an ESC/POS bitmap for use in the DSL}
\usage{
pos_plot(
  pos_obj,
  plot = ggplot2::last_plot(),
  color = FALSE,
  scale = 2,
  width = 256,
  height = 256,
  units = "px",
  dpi = 144,
  bg = "white",
  ...
)
}
\arguments{
\item{pos_obj}{object created with \code{\link[=escpos]{escpos()}}}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{color}{color?}

\item{scale, width, height, units, dpi, bg}{same as their \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} counterparts but with
sensible defaults for ESC/POS devices.}

\item{...}{other parameters to pass on to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}

\item{host_pos}{hostname or IP address of the ESC/POS compatible network device}

\item{port}{port the ESC/POS compatible device is listening on; defaults to \code{9100L}}
}
\description{
Turn a ggplot (or other grid object) intto an ESC/POS bitmap for use in the DSL
}
