% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{pos_size}
\alias{pos_size}
\title{Text size}
\usage{
pos_size(pos_obj, set = c("normal", "2h", "2w", "2x"))
}
\arguments{
\item{pos_obj}{object created with \code{\link[=escpos]{escpos()}}}

\item{set}{one of "\code{normal}", "\verb{2h}", "\verb{2w}", "\verb{2x}"}
}
\value{
\code{pos_obj} (invisibly)
}
\description{
Text size
}
