% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{pos_underline}
\alias{pos_underline}
\title{Underline}
\usage{
pos_underline(pos_obj, set = c("1dot", "2dot", "off"))
}
\arguments{
\item{pos_obj}{object created with \code{\link[=escpos]{escpos()}}}

\item{set}{one of "\verb{1dot}", "\verb{2dot}", "\code{off}"}
}
\value{
\code{pos_obj} (invisibly)
}
\description{
Underline
}
