% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R, R/escpos-package.R
\docType{package}
\name{escpos}
\alias{escpos}
\alias{escpos-package}
\title{Create an \code{escpos} object for accumulating print commands}
\usage{
escpos(host, port = 9100L)
}
\description{
ESC/POS devices, such as receipt printers, have the ability to print their own form of
raster bitmap graphics as well as text. Tools are provided to turn {ggplot2} or {grid} objects
into such raster bitmaps and print them, along with text and regular impages, to ESC/POS
compatible devices.#'
}
\examples{
escpos(epson_ip) |>
  pos_plaintext(
    stringi::stri_rand_lipsum(1)
  ) |>
  pos_lf(2) |>
  pos_plaintext(
    paste0(capture.output(
      str(mtcars, width = 40, strict.width = "cut")
    ), collapse = "\n")
  ) |>
  pos_lf(2L) |>
  pos_plaintext(
    stringi::stri_rand_lipsum(1, start_lipsum = FALSE)
  ) |>
  pos_lf(3) |>
  pos_cut() |>
  pos_print()
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/escpos}
  \item Report bugs at \url{https://github.com/hrbrmstr/escpos/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
