% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpos-package.R
\docType{package}
\name{ggpos}
\alias{ggpos}
\alias{ggpos-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/ggpos}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/ggpos/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
